* commit '8dbf353100bca0179367e07b45f169f1a823e023':
  Fix containers methods
  Fix deprecated methods
  Update version
  Mark as depricated legacy classes
  Mark as depricated legacy classes
  Fix publish to cocoapods
  Fix publish to cocoapods
  Fix publish to cocoapods
  Fix publish to cocoapods
  Restructuring project #2
  Restructuring project

# Conflicts:
#	Globus/CoreFoundation/Date/NSDate+GLBNS.h
#	Globus/CoreFoundation/Dictionary/NSDictionary+GLBNS.h
#	Globus/CoreFoundation/GLBTargetConditionals.h
#	Globus/CoreFoundation/NSAttributedString+GLBNS.h
#	Globus/CoreFoundation/NSDateFormatter+GLBNS.h
#	Globus/CoreFoundation/NSError+GLBNS.h
#	Globus/CoreFoundation/NSError+GLBNS.m
#	Globus/CoreFoundation/NSFileManager+GLBNS.h
#	Globus/CoreFoundation/NSHTTPCookieStorage+GLBNS.h
#	Globus/CoreFoundation/NSNull+GLBNS.m
#	Globus/CoreFoundation/NSNumber+GLBNS.m
#	Globus/CoreFoundation/NSPointerArray+GLBNS.m
#	Globus/CoreFoundation/NSURL+GLBNS.h
#	Globus/CoreFoundation/OrderedSet/NSOrderedSet+GLBNS.h
#	Globus/CoreFoundation/Pack/NSObject+GLBPack.h
#	Globus/CoreFoundation/Set/NSSet+GLBNS.h
#	Globus/CoreFoundation/Stream/NSStream+GLBNS.h
#	Globus/CoreFoundation/String/NSString+GLBNS.h

+++ Globus/AudioPlayer/GLBAudioPlayer.m
@@ -1,8 +1,6 @@
 /*--------------------------------------------------*/
 
 #import "GLBAudioPlayer.h"
-#import "GLBAudioSession.h"
-#import "GLBAction.h"
 
 /*--------------------------------------------------*/
 #if defined(GLB_TARGET_IOS)
@@ -69,7 +67,7 @@
 }
 
 - (void)setVolume:(CGFloat)volume {
-    _player.volume = volume;
+    _player.volume = (float)volume;
 }
 
 - (CGFloat)volume {
@@ -77,7 +75,7 @@
 }
 
 - (void)setPan:(CGFloat)pan {
-    _player.pan = pan;
+    _player.pan = (float)pan;
 }
 
 - (CGFloat)pan {
@@ -93,7 +91,7 @@
 }
 
 - (void)setRate:(CGFloat)rate {
-    _player.rate = rate;
+    _player.rate = (float)rate;
 }
 
 - (CGFloat)rate {
@@ -117,7 +115,7 @@
 }
 
 - (CGFloat)peakPower {
-    CGFloat result = 0.0f;
+    CGFloat result = 0.0;
     if(_prepared == YES) {
         NSUInteger numberOfChannels = _player.numberOfChannels;
         for(NSUInteger channel = 0; channel < numberOfChannels; channel++) {
@@ -129,7 +127,7 @@
 }
 
 - (CGFloat)averagePower {
-    CGFloat result = 0.0f;
+    CGFloat result = 0.0;
     if(_prepared == YES) {
         NSUInteger numberOfChannels = _player.numberOfChannels;
         for(NSUInteger channel = 0; channel < numberOfChannels; channel++) {
@@ -248,7 +246,7 @@
     if((_prepared == YES) && (_playing == YES)) {
         _playing = NO;
         _paused = NO;
-        _player.currentTime = 0.0f;
+        _player.currentTime = 0.0;
         [_player stop];
         if(_actionStoped != nil) {
             [_actionStoped performWithArguments:@[ self ]];
@@ -284,7 +282,7 @@
 }
 
 - (CGFloat)peakPowerForChannel:(NSUInteger)channelNumber {
-    CGFloat result = 0.0f;
+    CGFloat result = 0.0;
     if(_prepared == YES) {
         result = [_player peakPowerForChannel:channelNumber];
     }
@@ -292,7 +290,7 @@
 }
 
 - (CGFloat)averagePowerForChannel:(NSUInteger)channelNumber {
-    CGFloat result = 0.0f;
+    CGFloat result = 0.0;
     if(_prepared == YES) {
         result = [_player averagePowerForChannel:channelNumber];
     }
